+++ routes/connection.go
@@ -1,0 +1,30 @@
+package routes
+
+import (
+	"context"
+	"log"
+	"time"
+	"fmt"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/gridfs"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func ConnectMongoDB() {
+	var err error
+	clientOptions := options.Client().ApplyURI("mongodb+srv://challenge48hynov:cha!!enge48ynov@challenge48h.xgbo0vx.mongodb.net/?retryWrites=true&w=majority&appName=Challenge48h")
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	Mongoclient, err = mongo.Connect(context.Background(), clientOptions)
+	if err != nil {
+		log.Fatal(err)
+	}
+	err = Mongoclient.Ping(ctx, nil)
+	if err != nil {
+		log.Fatal("Impossible de se connecter à MongoDB")
+	}
+	Bucket, _ = gridfs.NewBucket(Mongoclient.Database("Challenge48h"), options.GridFSBucket().SetName("images"))
+	fmt.Println("Connexion MongoDB réussie")
+}
