+++ routes/getClient.go
@@ -1,0 +1,24 @@
+package routes
+
+import (
+	"time"
+	"context"
+	"github.com/gin-gonic/gin"
+
+	"net/http"
+	"go.mongodb.org/mongo-driver/bson"
+
+)
+
+func GetClient(c *gin.Context) {
+	collection := Mongoclient.Database("Challenge48h").Collection("Client")
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	cursor, _ := collection.Find(ctx, bson.M{})
+	defer cursor.Close(ctx)
+	var clients []Client
+	_ = cursor.All(ctx, &clients)
+	c.JSON(http.StatusOK, gin.H{
+		"message": clients,
+	})
+}
